Boss Sprites
The sprites for the boss have been implemented. The state machine fully works and has its own unique states utilizing the new sprites, which the boss unit in the levels can use. The test levels (Tommy, Zyn, and BossOnly) are the only levels impacted in this build, and have successfully changed the boss's sprites to utilize the new boss sprite sheet in the sprite renderer and animator components.